feat: Add validate_permissions function to utils.py
This commit adds the validate_permissions function to the utils.py file in the custom_auth app. The function validates a list of permissions provided in the request data. It checks if each permission has a required "model" field and at least one valid "action" field. It also verifies if the permissions exist in the database. If any errors are found, they are returned as a list of dictionaries. If no errors are found, the function returns a list of permission IDs. This function will be used in the assign_permission endpoint of the UserViewSet.

Note: The recent user commits and repository commits are not relevant to this specific code change.